requirements.txt
flask
python-dotenv
text-to-font
